Funcion asignar isAlive y estadística de aciertos al final del programa Por ahora el porcentaje de aciertos es del 67.61% A ver si se puede mejorar un poco podando alguna rama... ejem CLASE ENTERA.... ejem........

+++ lib/libstruct.h
@@ -10,8 +10,8 @@ typedef struct dat{
 	bool book5;
 	bool isMarried;
 	bool isNoble;
-	float numDeadRelations;
-	float popularity;
+	double numDeadRelations;
+	double popularity;
 	bool isAlive;
 }datos;
 typedef struct cuenta_datos{
